Install devDependencies by default (#519)
Fixes #431 

Many users getting started with Heroku try to run a build step with webpack, gulp, etc only to find that it's in their devDependencies and not installed by default on Heroku. They then have to figure out how to tell Heroku to install these. Defaulting to installing everything makes for a smoother onboarding.

Changes the default behavior to install `devDependencies`, run the build step, and then prune the `devDependencies` once the build step is complete.