Fix issue with old Node and metrics plugin
The resolves an issue where a user that is not using a version
of Node supported by the metrics plugin (not 8, 9, or 10) opts
in to the feature. Since the plugin for the appropriate version
of Node is copied into the slug at build time, but the metrics
url is set at runtime, we need to verify that the plugin exists
before including it.