Add metadata for which step of the build we are in (#677)
These steps get over-written each time, but if the build fails, we will be able to separate which stage it failed in.

A build that fails during the user-defined build-script stage is very different from a build that fails in the initial setup before user code is even executed. This should allow us to track those separately.