Use version resolver from heroku/nodejs CNB
The heroku/nodejs CNB uses a version resolver that is written in rust
and leverages a local inventory to resolve both Node.js and Yarn version
ranges to an available version. This will replace the Go version which
relied on querying the contents of an S3 bucket.

This change should reduce the dependency on http during builds. It also
unifies the heroku/nodejs CNB and v2 buildpacks to use the same
version resolution system.