Fix downgrade to npm 5.x when using version 10.x or greater
Several checks of the `major` version of npm being used were only accounting for the first character of the version string. E.g.; version `10.1.0` would be treated as `major=1` instead of `major=10`.

This PR changes those checks to split on `.` characters using `cut` and read the `major` version as the first field and adds a test to verify.

Fixes #1140