fix: do not ignore errors when checking npm/yarn output
Dropping stderr/stdout on the floor results in mysterious buildpack failures in certain cases. For instance, if there is a `.yarnrc.yml` that requires an environment variable, `yarn --version` will output an error, but then the existing code would just silence that, causing users to think that the yarn install failed with no real recourse.