Cleanup CHANGELOG.md (#1202)
* Cleanup CHANGELOG.md

This PR cleans up CHANGELOG.md to better match the format expected by the automation added in #1200. This will allow for release links to be created which we were not doing with our manual process.

* Cleanup CHANGELOG.md

This PR cleans up CHANGELOG.md to better match the format expected by the automation added in #1200. This will allow for release links to be created which we were not doing with our manual process.

* Update CHANGELOG.md

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

---------

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>