Classify peer dependency conflict error
- detects when `npm ci` is executed and a peer dependency conflict occurs
- provides an error message with appropriate detail and remediation steps
- classifies the error as `npm-peer-dependency-conflict`

A variant of this error which succeeds for `npm install` but fails for `npm prune` due to the absence of a `package-lock.json` is included. This is classified as `unknown-prune-dependencies-error` which is different from the above case. This classification will be used to identify a set of errors that aren't caught before pruning but should be.

- [W-19282675](https://gus.lightning.force.com/a07EE00002Jl7p1YAB)
- [W-19282801](https://gus.lightning.force.com/a07EE00002JlA5hYAF)