Remove corepack tooling to install package managers
Corepack will no longer be bundled with Node.js starting with the 25.0.0 release. For this buildpack the Corepack tooling is not actually required to install a package manager, it was just a convenient alternative. It's simpler to just remove the tooling and switch to installing the requested package managers using the npm registry.

This means developers can still choose to use Corepack tooling locally and this buildpack will respect their configuration via the `packageManager` field in `package.json`.

[Node.js TSC Vote](https://github.com/nodejs/TSC/pull/1697#issuecomment-2737093616)
[Node.js v25.0.0 Release Notes](https://github.com/nodejs/node/releases/tag/v25.0.0)
[Node.js v25.0.0 Corepack docs](https://nodejs.org/dist/latest/docs/api/corepack.html)